+++ src/umcm/__init__.py
@@ -2,4 +2,4 @@
 
 from umcm.ir import *  # noqa: F401,F403
 
-__version__ = "0.3.1"
+__version__ = "0.4.0"